Add addComment functionality
Adds functions to addComment.js to display a comment form
on click of the new comment button. Functions added to
take input and append a new comment to the comment_list
that exists on the html. The challenging pieces here were
finding the 'display: none' piece in the CSS and then also
figuring out how to undisplay/redisplay both the form and the
new comment button. There was also an issue that Nate was kind
enough to help with as far as wehre the listener for the
createcomment button needed to live in order to find the input
correctly.